Tidy up node post-processing dependencies.
The previous selection of packages had high impact security
issues reported. Looking into why postcss-cli v6 is considered
a breaking change from v5 it appears to be from
dropping node 4 and node 9 support, and old browsers from
client-side compiling, which we don't use as far as I know.
This upgrade makes two birds in one stone.

On top of which I've minimized the package.json since we're purely
using it for tracking devDependencies not actual packaging.