Migrate to docsy theme (#110)
* first draft of new theme is ready!

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* Tidy up node post-processing dependencies.

The previous selection of packages had high impact security
issues reported. Looking into why postcss-cli v6 is considered
a breaking change from v5 it appears to be from
dropping node 4 and node 9 support, and old browsers from
client-side compiling, which we don't use as far as I know.
This upgrade makes two birds in one stone.

On top of which I've minimized the package.json since we're purely
using it for tracking devDependencies not actual packaging.

* Disable searching for now, as it depends on Google and would mine data.

Whether we should or shouldn't use it or implement a local alternative
will be deferred to another issue / PR.

* Switch to git submodule tracking of the theme.

Docsy recommends this here
https://www.docsy.dev/docs/getting-started/#using-the-docsy-theme-as-a-submodule

Aside from easy updates, it keeps the history and licensing terms neatly separate.

* Expand on the exact dependencies we have using docsy.

* Suggest a different docker image compatible with docsy.

This has the required git and hugo extended dependencies.
However it doesn't have node, so it only works for previewing.

* Suggest installing hugo binary from a release over compiling from source.

* Add docsy reference in overview.

* Make the 404 page match other page styles.

* Adhere to the attribution line from LICENSE.md

* Use a Discord invite link instead of channel reference.

As a channel reference does not allow new people to join.

* Include bugfix in public to be able to recurse submodules.

* Add cleaning capability to the deploy script.

Especially because we're switching theme this is useful,
but it's also makes builds more reproducible as leftover assets
will be deleted this way.

* Update submodules in deploy.sh, because docsy is now a submodule.

* Refer to the updated klakegg/hugo and hugo version.

We're upgrading because this has all dependencies in one image.
See https://github.com/sfosc/sfosc/pull/110#issuecomment-518810641

* Because we clean the output, include static/CNAME

Based on documentation suggestion here:
https://gohugo.io/hosting-and-deployment/hosting-on-github/#use-a-custom-domain

* Use klakegg/hugo *-ext image to support new theme dependencies.

* Use the debian variety as it includes ssh and ssh-keygen.

Ubuntu would be an option as well. It's about a 5MB size difference.

* Explicitly ensure we have the right node dependencies.

* Explicitly check for node as well.

Although this was implied with npm, hugo will invoke node.
Checking both is less brittle.

* Use the short title to prevent overflows.

Resolves the issue raised here
https://github.com/sfosc/sfosc/pull/110#issuecomment-519180580